requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
tiktoken
langchain
langchain_community
pypdf
langchain_huggingface
